feat(nvim): add OverseerRestartLast command
This commit introduces a new Neovim command, `OverseerRestartLast`, which
restarts the most recent task in the Overseer task list. This command is
useful for quickly rerunning tasks without having to manually select them
from the list. The command is bound to the `<leader>rl` key combination.

In addition, the key binding for the `OverseerQuickAction` command has been
changed from `<leader>r` to `<leader>rr` to avoid conflicts.